Telegram bot/Food-Recipe-Telegram-Bot-using-OpenAI/requirements.txt
aiogram==2.15
openai==0.28
python-dotenv